make the search more robust when deleting a page, a content entry or a site

+++ app/jobs/locomotive/search_delete_content_entry_index_job.rb
@@ -1,11 +1,11 @@
 module Locomotive
-
   class SearchDeleteContentEntryIndexJob < BaseSearchJob
 
     def perform(site_id, content_type_slug, entry_id, locale)
       ::Mongoid::Fields::I18n.with_locale(locale) do
-        site  = Locomotive::Site.find(site_id)
-        entry = Locomotive::ContentEntry.find(entry_id)
+        site = Locomotive::Site.where(_id: site_id).first
+
+        return if site.nil? # usecase: the site has already been destroyed
 
         search_backend(site, locale)&.delete_object(
           content_type_slug,
@@ -15,5 +15,4 @@ module Locomotive
     end
 
   end
-
 end
